Move Operation Go structs from cluster-agent to backend-shared component, add CRD gen to backend-shared, fix DB schema bugs
