+++ backend-shared/.gitignore
@@ -1,3 +1,4 @@
+bin/
 dist/
 coverage.out
 vendor/
