+++ backend-shared/apis/managed-gitops/v1alpha1/operation_types.go
@@ -20,16 +20,9 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-// EDIT THIS FILE!  THIS IS SCAFFOLDING FOR YOU TO OWN!
-// NOTE: json tags are required.  Any new fields you add must have json tags for the fields to be serialized.
-
 // OperationSpec defines the desired state of Operation
 type OperationSpec struct {
-	// INSERT ADDITIONAL SPEC FIELDS - desired state of cluster
-	// Important: Run "make" to regenerate code after modifying this file
-
-	// Foo is an example field of Operation. Edit operation_types.go to remove/update
-	Foo string `json:"foo,omitempty"`
+	OperationID string `json:"operationID,omitempty"`
 }
 
 // OperationStatus defines the observed state of Operation
