+++ backend-shared/go.mod
@@ -7,9 +7,10 @@ require (
 	github.com/go-pg/pg/extra/pgdebug v0.2.0
 	github.com/go-pg/pg/v10 v10.10.6
 	github.com/google/uuid v1.3.0
-	github.com/onsi/ginkgo v1.16.4 // indirect
-	github.com/onsi/gomega v1.13.0 // indirect
 	github.com/stretchr/testify v1.7.0
 	golang.org/x/net v0.0.0-20210805182204-aaa1db679c0d // indirect
 	golang.org/x/sys v0.0.0-20211007075335-d3039528d8ac // indirect
+	k8s.io/apimachinery v0.22.4
+	k8s.io/code-generator v0.22.4 // indirect
+	sigs.k8s.io/controller-runtime v0.9.2
 )
