+++ cluster-agent/PROJECT
@@ -6,7 +6,7 @@ plugins:
   manifests.sdk.operatorframework.io/v2: {}
   scorecard.sdk.operatorframework.io/v2: {}
 projectName: cluster-agent
-repo: github.com/jgwest/managed-gitops/cluster-agent
+repo: github.com/redhat-appstudio/managed-gitops/cluster-agent
 resources:
 - api:
     crdVersion: v1
@@ -15,6 +15,6 @@ resources:
   domain: redhat.com
   group: managed-gitops
   kind: Operation
-  path: github.com/jgwest/managed-gitops/cluster-agent/api/v1alpha1
+  path: github.com/redhat-appstudio/managed-gitops/cluster-agent/apis/managed-gitops/v1alpha1
   version: v1alpha1
 version: "3"
