+++ cluster-agent/config/manifests/bases/cluster-agent.clusterserviceversion.yaml
@@ -1,0 +1,45 @@
+apiVersion: operators.coreos.com/v1alpha1
+kind: ClusterServiceVersion
+metadata:
+  annotations:
+    alm-examples: '[]'
+    capabilities: Basic Install
+  name: cluster-agent.v0.0.0
+  namespace: placeholder
+spec:
+  apiservicedefinitions: {}
+  customresourcedefinitions:
+    owned:
+    - description: Operation is the Schema for the operations API
+      displayName: Operation
+      kind: Operation
+      name: operations.managed-gitops.redhat.com
+      version: v1alpha1
+  description: b
+  displayName: a
+  icon:
+  - base64data: ""
+    mediatype: ""
+  install:
+    spec:
+      deployments: null
+    strategy: ""
+  installModes:
+  - supported: false
+    type: OwnNamespace
+  - supported: false
+    type: SingleNamespace
+  - supported: false
+    type: MultiNamespace
+  - supported: true
+    type: AllNamespaces
+  keywords:
+  - e
+  links:
+  - name: Cluster Agent
+    url: https://cluster-agent.domain
+  maturity: alpha
+  provider:
+    name: c
+    url: d
+  version: 0.0.0
